This page contain all of the exsiting API we can use for fractional derivative.

FractionalCalculus.fracdiff — Function

fracdiff(f::Function, α, start_point, end_point, ::FracDiffAlg)

Compute the α-order fractional derivative of f from start point to end point with specific algorithm.

fracdiff(f::Function, α, start_point, end_point, step_size ::Caputo)

Using Caputo definition to compute fractional derivative with complex step differentiation.

Example:

fracdiff(x->x^5, 0.5, 0, 2.5, 0.0001, Caputo_Direct())

Returns a tuple (result, error), which means the value of this derivative is 141.59714979764541, and the error estimate is 1.1532243848672914e-6.

When the input end points is an array, fracdiff will compute

https://en.wikipedia.org/wiki/Fractionalcalculus#Caputofractional_derivative

fracdiff(f, α, start_point, end_point, ::GL)

Using Grunwald-Letnikov definition to compute fractional derivative.

Example:

fracdiff(x->x^5, 0, 0.5, GL_Direct())

Please note Grunwald-Letnikov sense fracdiff only support 0 < α < 1.

Please refer to https://en.wikipedia.org/wiki/Gr%C3%BCnwald%E2%80%93Letnikov_derivative.

If the first order derivative of a function is already known,

We can use this method to compute the fractional order derivative more precisely.

If the first and second order derivative of a function is already known

Using the piecewise algorithm to obtain the fractional derivative at a specific point.

The RLDiff_Approx algorithm only support for 0 < α < 1