Fractional integral API

FractionalCalculus.jl has a disciplinary type system, by specifying which type you want to use, you can use the relating algorithm to compute fractional differentiation and fractional integral.

This page contains all of the existing API we can use for computing fractional integral.

FractionalCalculus.fracint — Function

fracint(f, α, point, h, FracIntAlg())

The general function used for fractional integral computing.

Example

julia> fracint(x->x^2, 0.5, 1, 0.001, RLInt_Approx())
0.6017877646029814

There are many algorithms can be used to computing fractional integral, please refer to our manual for more details: manual.

Riemann Liouville sense fractional integral

FractionalCalculus.RLInt — Type

Riemann-Liouville sense fractional integral algorithms

As the general fractional order integral type, RLInt is the parent type of all Riemann Liouville sense numerical methods.

Missing docstring.

Missing docstring for FractionalCalculus.RL_Direct. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FractionalCalculus.RL_Piecewise. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FractionalCalculus.RLInt_Approx. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FractionalCalculus.RL_LinearInterp. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FractionalCalculus.RLInt_Matrix. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FractionalCalculus.RLInt_Simpson. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FractionalCalculus.RLInt_Trapezoidal. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FractionalCalculus.RLInt_Rectangular. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FractionalCalculus.RLInt_Cubic_Spline_Interp. Check Documenter's build log for details.

Hadamard sense fractional integral

FractionalCalculus.HadamardMat

FractionalCalculus.@fracint — Macro

@fracint(f, α, point)

Return the α-order integral of $f$ at specific point.

julia> @fracint(x->x, 0.5, 1)
0.7522525439593486

FractionalCalculus.@semifracint — Macro

@semifracint(f, point)

Return the semi-integral of $f$ at specific point.

julia> @semifracint(x->x, 1)
0.7522525439593486

Symbolic fractional integral

FractionalCalculus.semiint — Function

Symbolic fractional integral

semiint(fun)

semiint uses SymbolicUtils.jl to compute symbolic fractional integral.

Example

julia> semiint(x^4)
0.45851597901024005(x^4.5)