Fractional derivative API

FractionalCalculus.jl has a disciplinary type system, by specifying which type you want to use, you can use the relating algorithm to compute fractional differentiation and fractional integral.

This page contains all of the existing API we can use for computing fractional derivative.

FractionalCalculus.fracdiff — Function

fracdiff(f, α, point, h, FracDiffAlg())

fracdiff is the general function for computing fractional derivative.

Example

julia> fracdiff(x->x^2, 0.5, 1, 0.001, RLDiff_Approx())
1.5044908143658473

There are many algorithms can be used to computing fractional derivative, please refer to our manual for more details: http://scifracx.org/FractionalCalculus.jl/dev/Derivative/derivativeapi/.

Caputo Sense fractional derivative

Missing docstring.

Missing docstring for FractionalCalculus.Caputo_Direct. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FractionalCalculus.Caputo_Piecewise. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FractionalCalculus.Caputo_Diethelm. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FractionalCalculus.Caputo_High_Precision. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FractionalCalculus.Caputo_High_Order. Check Documenter's build log for details.

Grunwald Letnikov sense fractional derivative

Missing docstring.

Missing docstring for FractionalCalculus.GL_Direct. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FractionalCalculus.GL_Finite_Difference. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FractionalCalculus.GL_Multiplicative_Additive. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FractionalCalculus.GL_Lagrange_Three_Point_Interp. Check Documenter's build log for details.

Riemann Liouville sense fractional derivative

Missing docstring.

Missing docstring for FractionalCalculus.RLDiff_Approx. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FractionalCalculus.RLDiff_Matrix. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FractionalCalculus.RL_G1. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FractionalCalculus.RL_D. Check Documenter's build log for details.

Hadamard sense fractional derivative

Missing docstring.

Missing docstring for FractionalCalculus.Hadamard_LRect. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FractionalCalculus.Hadamard_RRect. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FractionalCalculus.Hadamard_Trap. Check Documenter's build log for details.

Riesz sense fractional derivative

Missing docstring.

Missing docstring for FractionalCalculus.Riesz_Symmetric. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FractionalCalculus.Riesz_Ortigueira. Check Documenter's build log for details.

FractionalCalculus.@fracdiff — Macro

@fracdiff(f, α, point)

Return the α-order derivative of f at specific point.

julia> @fracdiff(x->x, 0.5, 1)
1.1283791670955188

FractionalCalculus.@semifracdiff — Macro

@semifracdiff(f, point)

Return the semi-derivative of f at spedific point.

julia> @semifracdiff(x->x, 1)
1.1283791670955188

Symbolic fractional differentiation

FractionalCalculus.semidiff — Function

Symbolic fractional differentiation

semidiff(fun)

semidiff uses SymbolicUtils.jl to compute symbolic fractional differentiation.

Example

julia> semidiff(log(x))
log(4x) / sqrt(πx)