Fractional derivative API

FractionalCalculus.jl has a disciplinary type system, by specifying which type you want to use, you can use the relating algorithm to compute fractional differentiation and fractional integral.

This page contain all of the existing API we can use for computing fractional derivative.

FractionalCalculus.fracdiff — Function

fracdiff(f, α, point, h, FracDiffAlg())

fracdiff is the general API for computing fractional derivative

Caputo Sense fractional derivative

FractionalCalculus.Caputo_Direct — Type

Caputo sense fractional derivative.

fracdiff(f::Function, α, start_point, end_point, step_size ::Caputo)

Example:

julia> fracdiff(x->x^5, 0.5, 0, 2.5, 0.0001, Caputo_Direct())

Returns a tuple (result, error), which means the value of this derivative is 141.59714979764541, and the error estimate is 1.1532243848672914e-6.

Refer to Caputo derivative

Note

Caputo Direct algorithms belong to direct computing, precise are ensured, but maybe cause more memory allocation and take more compilation time.

Using the direct mathematic expression:

\[^CD_t^α=\frac{1}{\Gamma(n-α)}\int_0^t\frac{f^{(n)}(τ)}{(t-τ)^{α+1-n}}\]

As for the derivative inside the integral, we use the Complex Step Differentiation to obtain a more accurate value.

FractionalCalculus.Caputo_Piecewise — Type

Caputo sense Piecewise algorithm

fracdiff(f, α, end_point, h, Caputo_Piecewise())

Using piecewise linear interpolation function to approximate input function and combining Caputo derivative then implement summation.

Example

julia> fracdiff(x->x^5, 0.5, 2.5, 0.001, Caputo_Piecewise())

Return the fractional derivative of $f(x)=x^5$ at point $x=2.5$.

@article{LI20113352,
title = {Numerical approaches to fractional calculus and fractional ordinary differential equation},
author = {Changpin Li and An Chen and Junjie Ye},
}

FractionalCalculus.Caputo_Diethelm — Type

Caputo sense Diethelm computation

fracdiff(f, α, end_point, h, Caputo_Diethelm())

Using quadrature weights(derived from product trapezoidal rule) to approximate the derivative.

Example

julia> fracdiff(x->x, 0.5, 1, 0.007, Caputo_Diethelm())
1.128378318687192

Scope

0 < α < 1

Diethelm's method for computingn Caputo sense fractional derivative using product trapezoidal rule and Richardsin extrapolation

@article{10.1093/imanum/17.3.479,
author = {DIETTELM, KAI},
title = "{Generalized compound quadrature formulae for finite-part integrals}",
journal = {IMA Journal of Numerical Analysis},
doi = {10.1093/imanum/17.3.479},
}

FractionalCalculus.Caputo_High_Precision — Type

Caputo sense fractioal derivative with p-th order precision.

fracdiff(f, α, t, Caputo_High_Precision())

Use the high precision algorithm to compute the Caputo sense fractional derivative.

Grunwald Letnikov sense fractional derivative

FractionalCalculus.GL_Direct — Type

Grünwald–Letnikov sense fractional dervivative.

fracdiff(f, α, start_point, end_point, GL_Direct())

Example:

julia> fracdiff(x->x^5, 0, 0.5, GL_Direct())

Scope

Please note Grunwald-Letnikov sense fracdiff only support $0 < \alpha < 1$.

Please refer to Grünwald–Letnikov derivative for more details.

Grunwald Letnikov direct compute method to obtain fractional derivative, precision are guaranteed but cause more memory allocation and compilation time.

FractionalCalculus.GL_Finite_Difference — Type

Grünwald Letnikov sense derivative approximation

fracdiff(f::Union{Function, Number}, α::AbstractArray, end_point, h, ::GL_Finite_Difference)::Vector

Use finite difference method to obtain Grünwald Letnikov sense fractional derivative.

Example

julia> fracdiff(x->x, 0.5, 1, 0.01, GL_Finite_Difference())
1.1269695801851276

Inaccurate

GL_Finite_Difference method is not accruate, please use it at your own risk.

FractionalCalculus.GL_Multiplicative_Additive — Type

Grünwald Letnikov sense derivative approximation

fracdiff(f, α, end_point, h, GL_Multiplicative_Additive())

Grünwald–Letnikov multiplication-addition-multiplication-addition··· method to approximate fractional derivative.

Example

julia> fracdiff(x->x, 0.5, 1, 0.007, GL_Multiplicative_Additive())
1.127403405642918

Inaccurate

The GL_Multiplicative_Additive method is not accruate, please use it at your own risk.

@book{oldham_spanier_1984,
title={The fractional calculus: Theory and applications of differentiation and integration to arbitrary order},
author={Oldham, Keith B. and Spanier, Jerome},
year={1984}} 

FractionalCalculus.GL_Lagrange_Three_Point_Interp — Type

Grünwald Letnikov sense derivative approximation

fracdiff(f, α, end_point, h, GL_Lagrange_Three_Point_Interp())

Using Lagrange three poitns interpolation to approximate the fractional derivative.

Example

julia> fracdiff(x->x, 0.5, 1, 0.007, GL_Lagrange_Three_Point_Interp())
1.1283963376811044

Inaccurate

The GL_Lagrange_Three_Point_Interp method is not accruate, please use it at your own risk.

Riemann Liouville sense fractional derivative

FractionalCalculus.RLDiff_Approx — Type

Riemann Liouville sense derivative approximation

fracdiff(f, α, end_point, h, RLDiff_Approx())

Using Linear interpolation to approximate fractional derivative in Riemann Liouville fractional derivative sense.

Example

julia> fracdiff(x->x^5, 0.5, 2.5, 0.0001, RLDiff_Approx())

Warning

The RLDiff_Approx algorithm only support for 0 < α < 1.

FractionalCalculus.RLDiff_Matrix — Type

Riemann Liouville sense derivative using Triangular Strip Matrix to discrete and compute.

fracdiff(f, α, end_point, h, RLDiff_Matrix())

Using Triangular Strip Matrix to approximate fractional derivative.

Example

julia> fracdiff(x->x^5, 0.5, 2.5, 0.0001, RLInt_Matrix())

Info

Triangular Strip Matrix method returns the derivative in the interval $[0, T]$ in Vector

Tip

With the advancing Triangular Strip Matrix method, you can not only compute fractional derivatives, integer order, higher order derivative is also supported!!

Try to set α as an integer, arbitrary integer of course! I promise you would enjoy it😏

@article{2009,
title={Matrix approach to discrete fractional calculus II: Partial fractional differential equations},
DOI={10.1016/j.jcp.2009.01.014},
author={Podlubny, Igor and Chechkin, Aleksei and Skovranek, Tomas and Chen, YangQuan and Vinagre Jara, Blas M.},
}

Hadamard sense fractional derivative

FractionalCalculus.Hadamard_LRect — Type

fracdiff(f, α, a, x, h, Hadamard_LRect())

Using finite part integral of left rectangular formula to compute the fractional hadamard derivative.

FractionalCalculus.Hadamard_RRect — Type

fracdiff(f, α, a, x, h, Hadamard_Trap())

Using finite part integral of right rectangular formula to compute the fractional hadamard derivative.

FractionalCalculus.Hadamard_Trap — Type

fracdiff(f, α, a, x, h, Hadamard_Trap())

Using finite part integral of trapezoidal formula to compute the fractional hadamard derivative.

Riesz sense fractional derivative

FractionalCalculus.Riesz_Symmetric — Type

Riesz sense symmetric fractional derivative algorithm.

fracdiff(f, α, end_point, h, Riesz_Symmetric())

Compute fractional derivative of Riesz sense using Triangular Strip Matrix algorithm.

Example

julia> fracdiff(x->x, 0.5, 1, 0.01, Riesz_Symmetric())

FractionalCalculus.@fracdiff — Macro

@fracdiff(f, α, point)

Return the α-order derivative of f at specific point.

julia> @fracdiff(x->x, 0.5, 1)
1.1283791670955188

FractionalCalculus.@semifracdiff — Macro

@semifracdiff(f, point)

Return the semi-derivative of f at spedific point.

julia> @semifracdiff(x->x, 1)
1.1283791670955188