This page contain all of the existing API we can use to compute fractional integral.

FractionalCalculus.fracint — Function

fracint as the general API used to compute fractional integral. By entering the function, the order, the start point and end point and step size, and choose the algorithms you want to use fracint(f::Function)

Example:

fracint(x->x^5, 0.5, 0, 2.5, 1e-8, RL_Direct())

Riemann_Liouville fractional integral using complex step differentiation Returns a tuple (1.1639316474512205, 1.0183453796725215e-8), which means the value of this derivative is 1.1639316474512205, and the error estimate is 1.0183453796725215e-8

With first order derivative known, we can directly use it in the computation of α order fraction integral

By deploying Piecewise interpolation to approximate the original function, with small step_size, this method is fast and take little memory allocation.

Note this algorithms demand α must be negative!!

RL_LinearInterp is more complex but more precise