Fractional integral API

FractionalCalculus.jl has a disciplinary type system, by specifying which type you want to use, you can use the relating algorithm to compute fractional differentiation and fractional integral.

This page contain all of the existing API we can use for computing fractional integral.

FractionalCalculus.fracint — Function

fracint(f, α, point, h, FracIntAlg())

The general API used for fractional integral computing.

Riemann Liouville sense fractional integral

FractionalCalculus.RLInt — Type

Riemann-Liouville sense fractional integral algorithms

Note this two algorithms belong to direct compute, precise are ensured, but maybe cause more memory allocation and take more compilation time.

FractionalCalculus.RL_Direct — Type

Riemann Liouville sense fractional integral

fracint(f::Function, α, start_point, end_point, h, RL_Direct())

Riemann_Liouville fractional integral using complex step differentiation.

\[(J^αf)(t)=\frac{1}{\Gamma(α)} \int_0^t(t-τ)^{α-1}f(τ)dτ\]

By using QuadGK calculate the integration and obtain the value.

Example:

julia> fracint(x->x^5, 0.5, 0, 2.5, 1e-8, RL_Direct())

Returns a tuple (1.1639316474512205, 1.0183453796725215e-8), which contains the value of this derivative is 1.1639316474512205, and the error estimate is 1.0183453796725215e-8

FractionalCalculus.RL_Piecewise — Type

Riemann Liouville sense fractional integral using piecewise interpolation.

fracint(f, α, end_point, h, RL_Piecewise())

Example

julia> fracint(x->x^5, 0.5, 2.5, 0.0001, RL_Piecewise())

By deploying Piecewise interpolation to approximate the original function, with small step size, this method is fast and take little memory allocation.

Using piecewise linear interpolation:

\[ y_n(t)=\frac{t-t_{i+1}}{t_i-t_{i+1}}y(t_i)+\frac{t-t_i}{t_{i+1}-t_i}y(t_{i+1})\]

Constitute the original function with the interpolation and implement the summation.

FractionalCalculus.RLInt_Approx — Type

Riemann Liouville sense fractional integral approximation.

fracint(f, α, end_point, h, RLInt_Approx())

Example

julia> fracint(x->x^5, 0.5, 2.5, 0.0001, RLInt_Approx())

Using the Staircase approximation to approximate the original function and implement the summation.

FractionalCalculus.RL_LinearInterp — Type

Riemann Liouville sense fractional integral using Linear interpolation.

fracint(f, α, end_point, h, RL_LinearInterp())

Example

julia> fracint(x->x^5, 0.5, 2.5, 0.0001, RL_LinearInterp())

RL_LinearInterp is more complex compared with RLInt_Approx but more precise.

Deploying the Linear Interpolation between $f_{j+1}$ and $f_j$, RL_LinearInterp method is more precise than RLInt_Approx.

FractionalCalculus.RLInt_Matrix — Type

Riemann Liouville sense integral using Triangular Strip Matrix to discrete.

fracint(f, α, end_point, h, RLInt_Matrix())

Using Triangular Strip Matrix to approximate fractional integral.

Example

julia> fracint(x->x^5, 0.5, 2.5, 0.0001, RLInt_Matrix())

Info

Triangular Strip Matrix method returns the derivative in the interval $[0, T]$ in Vector

Tip

With the advancing Triangular Strip Matrix method, you can not only compute fractional integrals, integer order, higher order integral is also supported!!

Try to set α as an integer, arbitrary integer of course! I promise you would enjoy it😏

Using Triangular Strip Matrix to discrete the integral.

FractionalCalculus.RLInt_Simpson — Type

Riemann Liouville sense fractional integral using fractional Simpson's formula to approximate.

Example

julia> fracint(x->x, 0.5, 1, 0.001, RLInt_Simpson)

@inproceedings{Li2015NumericalMF, title={Numerical Methods for Fractional Calculus}, author={Changpin Li and Fanhai Zeng}, year={2015} }

Using fractional Simpson's formula to discrete fractional integral.

FractionalCalculus.RLInt_Trapezoidal — Type

Riemann Liouville sense fractional integral using fractional trapezoidal formula to approximate.

Example

julia> fracint(x->x, 0.5, 1, 0.001, RLInt_Trapezoidal())

《Numerical methods for fractional calculus》. Using Trapezoidal method to approximate fractional integral

FractionalCalculus.RLInt_Rectangular — Type

Riemann Liouville sense fractional integral using fractional rectangular formula to approximate.

Example

julia> fracint(x->x, 0.5, 1, 0.001, RLInt_Rectangular())

FractionalCalculus.RLInt_Cubic_Spline_Interp — Type

Riemann Liouville sense fractional integral using cubic spline interpolation to approximate.

Example

julia> fracint(x->x, 0.5, 1, 0.0001, RLInt_Cubic_Spline_Interp())

Error estimate of this method is $\mathcal{O(h^4)}$, it is determined by the error of the cubic spline interpolation.

Set h as 0.001 or bigger

For some reason, in the RLIntCubicSpline_Interp method, set h as 0.001 would get better result.

Numerical Methods for Fractional Calculus Page 34

FractionalCalculus.@fracint — Macro

@fracint(f, α, point)

Return the α-order integral of f at specific point.

julia> @fracint(x->x, 0.5, 1)
0.7522525439593486

FractionalCalculus.@semifracint — Macro

@semifracint(f, point)

Return the semi-integral of f at specific point.

julia> @semifracint(x->x, 1)
0.7522525439593486