Fractional integral API

This page contain all of the existing API we can use to compute fractional integral.

FractionalCalculus.fracint — Function

fracint as the general API used to compute fractional integral. By entering the function, the order, the start point and end point and step size, and choose the algorithms you want to use fracint(f::Function)

Example:

fracint(x->x^5, 0.5, 0, 2.5, 1e-8, RL_Direct())

Riemann_Liouville fractional integral using complex step differentiation Returns a tuple (1.1639316474512205, 1.0183453796725215e-8), which means the value of this derivative is 1.1639316474512205, and the error estimate is 1.0183453796725215e-8

With first order derivative known, we can directly use it in the computation of α order fraction integral

By deploying Piecewise interpolation to approximate the original function, with small step_size, this method is fast and take little memory allocation.

Note this algorithms demand α must be negative!!

RL_LinearInterp is more complex but more precise

FractionalCalculus.FracIntAlg — Type

The base type of a fractional integral sense, in FractionalCalculus.jl, all of the fractional integral senses belong to FracIntAlg

FractionalCalculus.RLInt — Type

Riemann-Liouville sense fractional integral algorithms

Note this two algorithms belong to direct compute, precise are ensured, but maybe cause more memory allocation and take more compilation time.

FractionalCalculus.RL_Direct — Type

FractionalCalculus.RL_Direct_First_Diff_Known — Type

FractionalCalculus.RL_Piecewise — Type

Using piecewise linear interpolation:

\[ y_n(t)=\frac{t-t_{i+1}}{t_i-t_{i+1}}y(t_i)+\frac{t-t_i}{t_{i+1}-t_i}y(t_{i+1})\]

Constitute the original function with the interpolation and implement the summation.

FractionalCalculus.RLInt_Approx — Type

Using the Staircase approximation to approximate the original function and implement the summation.

Missing docstring.

Missing docstring for FractionalCalculus.RL_. Check Documenter's build log for details.